requirements.txt
django==3.2
django-cors-headers==3.12.0
djangorestframework==3.12.2
gunicorn==20.0.4
requests==2.19.1
aiohttp==3.8.1
psycopg2-binary==2.8.2
pyyaml==5.2